Rename LightingControlCard to LEDControlCard
