+++ photon-client/src/components/settings/LEDControlCard.vue
